--- a/app/views/home/styleguide.html.haml
+++ b/app/views/home/styleguide.html.haml
@@ -10,3 +10,11 @@
       = t("home.styleguide.swatch.work")
     .swatch.colors.work-completed
       = t("home.styleguide.swatch.work-completed")
+%section#icons
+  %h2= t(".icons.title")
+  %ul
+    %li= t(".icons.ready_html", { icon: t("icons.ready_html") })
+    %li= t(".icons.action_html", { icon: t("icons.arrow_html") })
+    %li= t(".icons.completed_html", { icon: t("icons.completed_html") })
+    %li= t(".icons.delete_html", { icon: t("icons.delete_html") })
+