--- a/app/views/layouts/application.html.haml
+++ b/app/views/layouts/application.html.haml
@@ -6,8 +6,6 @@
     %link{ href: "https://fonts.googleapis.com/css?family=Oswald:400,700", rel: "stylesheet", type: "text/css" }
     %link{ href: "https://fonts.googleapis.com/css?family=Open+Sans+Condensed:300", rel: "stylesheet", type: "text/css" }
     %link{ href: "https://fonts.googleapis.com/css?family=Special+Elite", rel: "stylesheet", type: "text/css" }
-    %link{ href: "https://maxcdn.bootstrapcdn.com/bootstrap/3.3.5/css/bootstrap.min.css", :rel => "stylesheet"}
-    %link{:href => "https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.3.0/css/font-awesome.min.css", :rel => "stylesheet"}
 
     %link{ href: asset_path("favicon.png"), type: "image/png", rel: :icon }
     %link{ rel: "apple-touch-icon", href: asset_path("icon.png") }