--- a/app/views/missions/show.html.haml
+++ b/app/views/missions/show.html.haml
@@ -1,11 +1,12 @@
-- if !@mission.estimate.zero?
+- progress = @mission.status
+- total_time = progress.values.sum
+- if total_time > 0
   .progress
-    .progress-bar.progress-bar-success.progress-bar-striped.active{:role => "progressbar", :style => "width:#{@mission.progress["completed"]}%"}
-      Completed - #{@mission.progress["completed"]}%
-    .progress-bar.progress-bar-warning.progress-bar-striped.active{:role => "progressbar", :style => "width:#{@mission.progress["started"]}%"}
-      In progress - #{@mission.progress["started"]}%
-    .progress-bar.progress-bar-danger.progress-bar-striped.active{:role => "progressbar", :style => "width:#{@mission.progress["created"]}%"}
-      Yet to start - #{@mission.progress["created"]}%
+    - progress.each do |(status, time)|
+      - percent = (time.fdiv(total_time) * 100)
+      %div{ class: "progress-#{status}" }
+        .progress-description= status
+        .progress-bar{ class: status, "data-percent" => percent  }
 
 %section
   .right= link_to(t("actions.edit"), edit_mission_path(@mission))
@@ -18,24 +19,26 @@
 
   - if @mission.deliverables.any?
     %ul#deliverables{ "data-id" => @mission.id }
-      - @mission.deliverables.reverse_order.each do |deliverable|
+      - @mission.deliverables.each do |deliverable|
         %li{ id: deliverable.to_param, "data-id" => deliverable.id }
           %span.deliverable_handle= t("icons.handle_html")
-          = link_to(t("missions.show.deliverable_html", { name: deliverable.name , estimate: deliverable.estimate}), edit_mission_deliverable_path(@mission, deliverable))
+          = link_to(deliverable.name, edit_mission_deliverable_path(@mission, deliverable))
           %span.subnote= deliverable.value
           %ul.requirements{ "data-id" => deliverable.id }
-            - deliverable.requirements.reverse_order.each do |requirement|
+            - deliverable.requirements.each do |requirement|
               %li.requirement{ id: requirement.to_param, "data-id" => requirement.id }
                 %span.requirement_handle= t("icons.handle_html")
                 = link_to(t("deliverables.show.requirement_html", { name: requirement.name, estimate: requirement.estimate}),edit_mission_deliverable_requirement_path(@mission,deliverable, requirement))
                 - if requirement.created?
                   - path = mission_deliverable_requirement_path(@mission, deliverable, requirement, { requirement: { status: Requirement::STATUSES[:started] } })
-                  = link_to(t("icons.class_icon_html"), path, { method: :put, title: "Start" , class: "fa fa-play-circle"})
+                  - text = t(".requirement.start_html", { icon: t("icons.ready_html") })
+                  = link_to(text, path, { method: :put, title: "Start" })
                 - elsif requirement.started?
                   - path = mission_deliverable_requirement_path(@mission, deliverable, requirement, { requirement: { status: Requirement::STATUSES[:completed] } })
-                  = link_to(t("icons.class_icon_html"), path, method: :put, title: "Finish", class: "fa fa-dot-circle-o" )
-                - else
-                  %i.fa.fa-check
+                  - text = t(".requirement.finish_html", { icon: t("icons.actioning_html") })
+                  = link_to(text, path, method: :put, title: "Finish")
+                %br
+                = requirement.description
             = link_to(new_mission_deliverable_requirement_path(@mission, deliverable), { title: t("missions.show.add-requirement") }) do
               = t("missions.show.add-requirement_html", { arrow: t("icons.arrow_html") })
   - else