--- a/spec/unit/models/mission_spec.rb
+++ b/spec/unit/models/mission_spec.rb
@@ -8,4 +8,27 @@ RSpec.describe Mission, type: :model do
       expect(mission.errors[:name]).to_not be_empty
     end
   end
+
+  describe("#status") do
+    it "summarizes estimates of requirements" do
+      mission = Generator.mission!
+      deliverables = 2.times.collect do
+        Generator.deliverable!({ mission: mission })
+      end
+      requirements = 3.times.collect do
+        Generator.requirement!({
+          mission: mission,
+          deliverable: deliverables.sample,
+          estimate: rand(10),
+          status: Requirement.statuses.keys.sample
+        })
+      end
+
+      estimates = Requirement.statuses.each_with_object({}) do |(name, value), estimates|
+        estimates[name] = mission.requirements.where({ status: value }).sum(:estimate)
+      end
+      expect(mission.status).to eq(estimates)
+    end
+
+  end
 end