… spec/unit/models/requirement_spec.rb