Fixing some n+1 problems, removing assets calls, etc. As part of this, it also updates our status bar:

<img width="989" alt="screen shot 2015-09-19 at 11 10 15 pm" src="https://cloud.githubusercontent.com/assets/96204/9979182/b15a6a66-5f23-11e5-97eb-8244956cecb1.png">


The missions#show page still isn't great:

```
Started GET "/missions/00fa3307-5bfa-4ec9-99d5-fbbb53598fea" for ::1 at 2015-09-19 22:44:32 -0500
Processing by MissionsController#show as HTML
Parameters: {"id"=>"00fa3307-5bfa-4ec9-99d5-fbbb53598fea"}
Mission Load (0.4ms)  SELECT  "missions".* FROM "missions" WHERE "missions"."id" = $1 LIMIT 1  [["id", "00fa3307-5bfa-4ec9-99d5-fbbb53598fea"]]
User Load (0.5ms)  SELECT  "users".* FROM "users" WHERE "users"."id" = $1 LIMIT 1  [["id", "410c2df1-de3c-43b9-9a39-76f0a70c9d0a"]]
User Exists (0.5ms)  SELECT  1 AS one FROM "users" INNER JOIN "participants" ON "users"."id" = "participants"."user_id" WHERE "participants"."joinable_id" = $1 AND "participants"."joinable_type" = $2 AND "users"."id" = $3 LIMIT 1  [["joinable_id", "00fa3307-5bfa-4ec9-99d5-fbbb53598fea"], ["joinable_type", "Mission"], ["id", "410c2df1-de3c-43b9-9a39-76f0a70c9d0a"]]
User Load (0.5ms)  SELECT "users".* FROM "users" INNER JOIN "participants" ON "users"."id" = "participants"."user_id" WHERE "participants"."joinable_id" = $1 AND "participants"."joinable_type" = $2  [["joinable_id", "00fa3307-5bfa-4ec9-99d5-fbbb53598fea"], ["joinable_type", "Mission"]]
Mission Load (0.5ms)  SELECT  "missions".* FROM "missions" WHERE "missions"."id" = $1  ORDER BY "missions"."id" ASC LIMIT 1  [["id", "00fa3307-5bfa-4ec9-99d5-fbbb53598fea"]]
Deliverable Load (0.6ms)  SELECT "deliverables".* FROM "deliverables" WHERE "deliverables"."mission_id" IN ('00fa3307-5bfa-4ec9-99d5-fbbb53598fea')  ORDER BY "deliverables"."ordering" ASC
Requirement Load (0.8ms)  SELECT "requirements".* FROM "requirements" WHERE "requirements"."deliverable_id" IN ('d1f6e2c6-76ec-4c57-98ec-d9db47964f3b', '21fa231d-0127-40a2-81b9-dc5133b8f887', '1eed5477-8425-4841-8339-32b9de0bd401')  ORDER BY "requirements"."ordering" ASC
Requirement Load (0.5ms)  SELECT SUM(estimate) as time, "requirements"."status" FROM "requirements" INNER JOIN "deliverables" ON "deliverables"."id" = "requirements"."deliverable_id" WHERE "deliverables"."mission_id" = $1 GROUP BY "requirements"."status"  [["mission_id", "00fa3307-5bfa-4ec9-99d5-fbbb53598fea"]]
Participant Load (0.5ms)  SELECT "participants".* FROM "participants" WHERE "participants"."joinable_id" = $1 AND "participants"."joinable_type" = $2  [["joinable_id", "00fa3307-5bfa-4ec9-99d5-fbbb53598fea"], ["joinable_type", "Mission"]]
Rendered missions/show.html.haml within layouts/application (328.4ms)
CACHE (0.2ms)  SELECT  "users".* FROM "users" WHERE "users"."id" = $1 LIMIT 1  [["id", "410c2df1-de3c-43b9-9a39-76f0a70c9d0a"]]
CACHE (0.2ms)  SELECT  "users".* FROM "users" WHERE "users"."id" = $1 LIMIT 1  [["id", "410c2df1-de3c-43b9-9a39-76f0a70c9d0a"]]
```

but it's certainly better than

```
ActiveRecord::SchemaMigration Load (0.4ms)  SELECT "schema_migrations".* FROM "schema_migrations"
Processing by MissionsController#show as HTML
Parameters: {"id"=>"00fa3307-5bfa-4ec9-99d5-fbbb53598fea"}
Mission Load (0.4ms)  SELECT  "missions".* FROM "missions" WHERE "missions"."id" = $1 LIMIT 1  [["id", "00fa3307-5bfa-4ec9-99d5-fbbb53598fea"]]
User Load (0.4ms)  SELECT  "users".* FROM "users" WHERE "users"."id" = $1 LIMIT 1  [["id", "410c2df1-de3c-43b9-9a39-76f0a70c9d0a"]]
User Exists (0.5ms)  SELECT  1 AS one FROM "users" INNER JOIN "participants" ON "users"."id" = "participants"."user_id" WHERE "participants"."joinable_id" = $1 AND "participants"."joinable_type" = $2 AND "users"."id" = $3 LIMIT 1  [["joinable_id", "00fa3307-5bfa-4ec9-99d5-fbbb53598fea"], ["joinable_type", "Mission"], ["id", "410c2df1-de3c-43b9-9a39-76f0a70c9d0a"]]
User Load (0.3ms)  SELECT "users".* FROM "users" INNER JOIN "participants" ON "users"."id" = "participants"."user_id" WHERE "participants"."joinable_id" = $1 AND "participants"."joinable_type" = $2  [["joinable_id", "00fa3307-5bfa-4ec9-99d5-fbbb53598fea"], ["joinable_type", "Mission"]]
Deliverable Load (0.5ms)  SELECT "deliverables".* FROM "deliverables" WHERE "deliverables"."mission_id" = $1  ORDER BY ordering DESC  [["mission_id", "00fa3307-5bfa-4ec9-99d5-fbbb53598fea"]]
(0.4ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "1eed5477-8425-4841-8339-32b9de0bd401"]]
(0.3ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "21fa231d-0127-40a2-81b9-dc5133b8f887"]]
(0.2ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "d1f6e2c6-76ec-4c57-98ec-d9db47964f3b"]]
Requirement Load (0.4ms)  SELECT "requirements".* FROM "requirements" WHERE "requirements"."deliverable_id" = $1  ORDER BY ordering DESC  [["deliverable_id", "1eed5477-8425-4841-8339-32b9de0bd401"]]
Requirement Load (0.4ms)  SELECT "requirements".* FROM "requirements" WHERE "requirements"."deliverable_id" = $1  ORDER BY ordering DESC  [["deliverable_id", "21fa231d-0127-40a2-81b9-dc5133b8f887"]]
Requirement Load (0.4ms)  SELECT "requirements".* FROM "requirements" WHERE "requirements"."deliverable_id" = $1  ORDER BY ordering DESC  [["deliverable_id", "d1f6e2c6-76ec-4c57-98ec-d9db47964f3b"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "1eed5477-8425-4841-8339-32b9de0bd401"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "21fa231d-0127-40a2-81b9-dc5133b8f887"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "d1f6e2c6-76ec-4c57-98ec-d9db47964f3b"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "1eed5477-8425-4841-8339-32b9de0bd401"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "21fa231d-0127-40a2-81b9-dc5133b8f887"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "d1f6e2c6-76ec-4c57-98ec-d9db47964f3b"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "1eed5477-8425-4841-8339-32b9de0bd401"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "21fa231d-0127-40a2-81b9-dc5133b8f887"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "d1f6e2c6-76ec-4c57-98ec-d9db47964f3b"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "1eed5477-8425-4841-8339-32b9de0bd401"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "21fa231d-0127-40a2-81b9-dc5133b8f887"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "d1f6e2c6-76ec-4c57-98ec-d9db47964f3b"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "1eed5477-8425-4841-8339-32b9de0bd401"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "21fa231d-0127-40a2-81b9-dc5133b8f887"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "d1f6e2c6-76ec-4c57-98ec-d9db47964f3b"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "1eed5477-8425-4841-8339-32b9de0bd401"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "21fa231d-0127-40a2-81b9-dc5133b8f887"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "d1f6e2c6-76ec-4c57-98ec-d9db47964f3b"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "1eed5477-8425-4841-8339-32b9de0bd401"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "21fa231d-0127-40a2-81b9-dc5133b8f887"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "d1f6e2c6-76ec-4c57-98ec-d9db47964f3b"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "1eed5477-8425-4841-8339-32b9de0bd401"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "21fa231d-0127-40a2-81b9-dc5133b8f887"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "d1f6e2c6-76ec-4c57-98ec-d9db47964f3b"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "1eed5477-8425-4841-8339-32b9de0bd401"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "21fa231d-0127-40a2-81b9-dc5133b8f887"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "d1f6e2c6-76ec-4c57-98ec-d9db47964f3b"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "1eed5477-8425-4841-8339-32b9de0bd401"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "21fa231d-0127-40a2-81b9-dc5133b8f887"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "d1f6e2c6-76ec-4c57-98ec-d9db47964f3b"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "1eed5477-8425-4841-8339-32b9de0bd401"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "21fa231d-0127-40a2-81b9-dc5133b8f887"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "d1f6e2c6-76ec-4c57-98ec-d9db47964f3b"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "1eed5477-8425-4841-8339-32b9de0bd401"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "21fa231d-0127-40a2-81b9-dc5133b8f887"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "d1f6e2c6-76ec-4c57-98ec-d9db47964f3b"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "1eed5477-8425-4841-8339-32b9de0bd401"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "21fa231d-0127-40a2-81b9-dc5133b8f887"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "d1f6e2c6-76ec-4c57-98ec-d9db47964f3b"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "1eed5477-8425-4841-8339-32b9de0bd401"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "21fa231d-0127-40a2-81b9-dc5133b8f887"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "d1f6e2c6-76ec-4c57-98ec-d9db47964f3b"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "1eed5477-8425-4841-8339-32b9de0bd401"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "21fa231d-0127-40a2-81b9-dc5133b8f887"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "d1f6e2c6-76ec-4c57-98ec-d9db47964f3b"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "1eed5477-8425-4841-8339-32b9de0bd401"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "21fa231d-0127-40a2-81b9-dc5133b8f887"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "d1f6e2c6-76ec-4c57-98ec-d9db47964f3b"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "1eed5477-8425-4841-8339-32b9de0bd401"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "21fa231d-0127-40a2-81b9-dc5133b8f887"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "d1f6e2c6-76ec-4c57-98ec-d9db47964f3b"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "1eed5477-8425-4841-8339-32b9de0bd401"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "21fa231d-0127-40a2-81b9-dc5133b8f887"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "d1f6e2c6-76ec-4c57-98ec-d9db47964f3b"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "1eed5477-8425-4841-8339-32b9de0bd401"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "21fa231d-0127-40a2-81b9-dc5133b8f887"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "d1f6e2c6-76ec-4c57-98ec-d9db47964f3b"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "1eed5477-8425-4841-8339-32b9de0bd401"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "21fa231d-0127-40a2-81b9-dc5133b8f887"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "d1f6e2c6-76ec-4c57-98ec-d9db47964f3b"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "1eed5477-8425-4841-8339-32b9de0bd401"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "21fa231d-0127-40a2-81b9-dc5133b8f887"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "d1f6e2c6-76ec-4c57-98ec-d9db47964f3b"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "1eed5477-8425-4841-8339-32b9de0bd401"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "21fa231d-0127-40a2-81b9-dc5133b8f887"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "d1f6e2c6-76ec-4c57-98ec-d9db47964f3b"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "1eed5477-8425-4841-8339-32b9de0bd401"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "21fa231d-0127-40a2-81b9-dc5133b8f887"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "d1f6e2c6-76ec-4c57-98ec-d9db47964f3b"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "1eed5477-8425-4841-8339-32b9de0bd401"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "21fa231d-0127-40a2-81b9-dc5133b8f887"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "d1f6e2c6-76ec-4c57-98ec-d9db47964f3b"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "1eed5477-8425-4841-8339-32b9de0bd401"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "21fa231d-0127-40a2-81b9-dc5133b8f887"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "d1f6e2c6-76ec-4c57-98ec-d9db47964f3b"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "1eed5477-8425-4841-8339-32b9de0bd401"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "21fa231d-0127-40a2-81b9-dc5133b8f887"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "d1f6e2c6-76ec-4c57-98ec-d9db47964f3b"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "1eed5477-8425-4841-8339-32b9de0bd401"]]
CACHE (0.1ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "21fa231d-0127-40a2-81b9-dc5133b8f887"]]
CACHE (0.0ms)  SELECT SUM("requirements"."estimate") FROM "requirements" WHERE "requirements"."deliverable_id" = $1  [["deliverable_id", "d1f6e2c6-76ec-4c57-98ec-d9db47964f3b"]]
…